SonarQube improvements

+++ src/Orc.LicenseManager.Client/Helpers/NetworkValidationHelper.cs
@@ -27,33 +27,30 @@ namespace Orc.LicenseManager
             }
 
             var validationResult = e.ValidationResult;
-            if (!validationResult.IsValid)
+            if (!validationResult.IsValid && validationResult.IsCurrentUserLatestUser())
             {
-                if (validationResult.IsCurrentUserLatestUser())
+                _isInErrorHandling = true;
+
+                var entryAssembly = AssemblyHelper.GetEntryAssembly();
+
+                var message = string.Format("The current number of usages for {0} is higher than the maximum number of concurrent users allowed based on the current license. Since this computer is the last one using the license, the software has to shut down.\n\nIf you feel that you have not reached the maximum number of usages, please contact support.\n\nThe maximum allowed is {1}, the current usage is {2}.", entryAssembly.Title(), validationResult.MaximumConcurrentUsers, validationResult.CurrentUsers.Count);
+
+                Log.Error(message);
+
+                Log.Error("Listing all the usages of the license:");
+                Log.Indent();
+
+                foreach (var licenseUsage in validationResult.CurrentUsers)
                 {
-                    _isInErrorHandling = true;
-
-                    var entryAssembly = AssemblyHelper.GetEntryAssembly();
-
-                    var message = string.Format("The current number of usages for {0} is higher than the maximum number of concurrent users allowed based on the current license. Since this computer is the last one using the license, the software has to shut down.\n\nIf you feel that you have not reached the maximum number of usages, please contact support.\n\nThe maximum allowed is {1}, the current usage is {2}.", entryAssembly.Title(), validationResult.MaximumConcurrentUsers, validationResult.CurrentUsers.Count);
-
-                    Log.Error(message);
-
-                    Log.Error("Listing all the usages of the license:");
-                    Log.Indent();
-
-                    foreach (var licenseUsage in validationResult.CurrentUsers)
-                    {
-                        Log.Error("* {0}", licenseUsage);
-                    }
-
-                    Log.Unindent();
-
-                    Log.Info("Shutting down application");
-
-                    var process = Process.GetCurrentProcess();
-                    process.Kill();
+                    Log.Error("* {0}", licenseUsage);
                 }
+
+                Log.Unindent();
+
+                Log.Info("Shutting down application");
+
+                var process = Process.GetCurrentProcess();
+                process.Kill();
             }
         }
     }
